+++ client/src/game/animal.h
@@ -7,6 +7,6 @@ extern uint animalCount;
 
 void animalDrawAll        ();
 void animalUpdateAll      ();
-int  animalHitCheck       (const vec pos, float mdd, int dmg, int cause, uint iteration);
+int  animalHitCheck       (const vec pos, float mdd, int dmg, int cause, u16 iteration);
 void animalSyncFromServer (const packet *p);
 void animalGotHitPacket   (const packet *p);
