+++ client/src/game/character.h
@@ -27,7 +27,7 @@ int   characterGetPlayerHP     (uint i);
 vec   characterGetPlayerDist   (uint i);
 void  characterGotHitPacket    (const packet *p);
 void  characterRemovePlayer    (int c, int len);
-int   characterHitCheck        (const vec pos, float mdd, int damage, int cause, uint iteration);
+int   characterHitCheck        (const vec pos, float mdd, int damage, int cause, u16 iteration);
 
 void  charactersUpdate         ();
 void  characterDrawAll         ();
